moves refactored conversation comment components to their own folder
